# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: SaaS Onboarding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials
...
  /admin/tenants:
    get:
      operationId: listTenants
      tags: [admin]
      summary: List all tenants
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [admin]
      summary: Create a new tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /admin/tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [admin]
      summary: Get tenant by ID
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    patch:
      operationId: updateTenant
      tags: [admin]
      summary: Update tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
    delete:
      operationId: deleteTenant
      tags: [admin]
      summary: Delete tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /admin/tenants/{tenantId}/analytics:
    get:
      operationId: getTenantAnalytics
      tags: [admin]
      summary: Get analytics for a specific tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tenant analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantAnalytics"
        "404":
          description: Not found
  /admin/tenants/{tenantId}/onboarding:
    get:
      operationId: listOnboardingSubmissions
      tags: [admin]
      summary: List onboarding submissions for a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Submissions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OnboardingSubmission"
  /admin/onboarding/{submissionId}:
    patch:
      operationId: updateOnboardingSubmission
      tags: [admin]
      summary: Update onboarding submission status
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingStatusUpdate"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingSubmission"
        "404":
          description: Not found
...
  /auth/business-login:
    post:
      operationId: businessLogin
      tags: [auth]
      summary: Business owner login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessLoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessToken"
        "401":
          description: Invalid credentials
  /auth/business-me:
    get:
      operationId: getBusinessMe
      tags: [auth]
      summary: Get current business user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessUser"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    AuthToken:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
      required:
        - token
        - expiresAt
    LoginInput:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        subdomain:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        plan:
          type: string
          enum: [free, starter, pro, enterprise]
        status:
          type: string
          enum: [active, inactive, suspended]
        whatsappPhoneNumberId:
          type: ["string", "null"]
        businessType:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        openingHours:
          type: ["string", "null"]
        trialEndsAt:
          type: ["string", "null"]
        onboardingStep:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - slug
        - subdomain
        - email
        - plan
        - status
        - createdAt
        - updatedAt
    TenantInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: string
        plan:
          type: string
          enum: [free, starter, pro, enterprise]
        whatsappPhoneNumberId:
          type: string
      required:
        - name
        - email
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        plan:
          type: string
          enum: [free, starter, pro, enterprise]
        status:
          type: string
          enum: [active, inactive, suspended]
        whatsappPhoneNumberId:
          type: string
...
    OnboardingSubmission:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        businessName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, reviewed, approved, rejected]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tenantId
        - businessName
        - contactName
        - email
        - phone
        - status
        - createdAt
        - updatedAt
...
    OnboardingStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, reviewed, approved, rejected]
      required:
        - status
...
    TenantAnalytics:
      type: object
      properties:
        tenantId:
          type: string
        onboardingCount:
          type: integer
        whatsappSessionCount:
          type: integer
        whatsappMessageCount:
          type: integer
        pendingSubmissions:
          type: integer
        approvedSubmissions:
          type: integer
      required:
        - tenantId
        - onboardingCount
        - whatsappSessionCount
        - whatsappMessageCount
        - pendingSubmissions
        - approvedSubmissions
...
    BusinessUser:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required:
        - id
        - tenantId
        - name
        - email
        - role
        - createdAt
...
    BusinessLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    BusinessToken:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
        user:
          $ref: "#/components/schemas/BusinessUser"
      required:
        - token
        - expiresAt
        - user
...